Use hardware AES rounds on supported compiler targets

+++ aes.h
@@ -26,6 +26,8 @@
 #ifndef POLARSSL_AES_H
 #define POLARSSL_AES_H
 
+#include <stdint.h>
+
 #define AES_ENCRYPT     1
 #define AES_DECRYPT     0
 
@@ -38,8 +40,8 @@
 typedef struct
 {
     int nr;                     /*!<  number of rounds  */
-    unsigned long *rk;          /*!<  AES round keys    */
-    unsigned long buf[68];      /*!<  unaligned data    */
+    uint32_t *rk;          /*!<  AES round keys    */
+    uint32_t buf[68];      /*!<  unaligned data    */
 }
 aes_context;
